backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0